fix: longitude, latitude Long -> Double로 변경

+++ hub/src/main/java/com/klp/hub/hub/presentation/dto/response/hub/GetHubListResponse.java
@@ -12,8 +12,8 @@ public record GetHubListResponse(
     public record HubSummaryResponse(
         UUID hubId,
         String name,
-        Long latitude,
-        Long longitude,
+        Double latitude,
+        Double longitude,
         String address
     ){}
